refactor: rename buttons/test.tsx to toggleable_button.tsx
